requirements.txt
Flask==2.3.3
pandas==2.0.3
numpy==1.26.4
openpyxl==3.1.2
qrcode==7.4.2
flask-cors==3.0.10
gunicorn==21.2.0
gspread==6.2.1
google-auth==2.21.0
google-auth-oauthlib==1.3.0
python-dotenv==1.0.1
face-recognition==1.3.0
dlib==19.24.1
Pillow==10.1.0